refactor(control-plane): clarify provider composition name
