+++ lambdas/functions/control-plane/src/test-framework/runner-provider/provider-types.ts
@@ -1,3 +1,3 @@
-import { runnerProviderTypes } from '../../provider-registry';
+import { runnerProviderTypes } from '../../control-plane-providers';
 
 export const providerTypes = runnerProviderTypes;
